feat: external module for popUp
